+++ GlobalSuppressions.cs
@@ -18,9 +18,9 @@
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.ReadabilityRules", "SA1124:Do not use regions", Justification = "Disagree in this case", Scope = "member", Target = "~M:Microsoft.Toolkit.Uwp.UI.Lottie.LottieVisualSource.CheckedAwaitAsync(System.Threading.Tasks.Task)~System.Threading.Tasks.Task")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.SpacingRules", "SA1001:Commas must be spaced correctly", Justification = "No space because it's a string format specifier", Scope = "member", Target = "~M:Program.WriteCodeGenStatsReport(System.IO.TextWriter,Microsoft.Toolkit.Uwp.UI.Lottie.WinCompData.Tools.Stats,Microsoft.Toolkit.Uwp.UI.Lottie.WinCompData.Tools.Stats)")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.SpacingRules", "SA1001:Commas must be spaced correctly", Justification = "No space because it's a string format specifier", Scope = "member", Target = "~M:Program.WriteLottieStatsReport(System.IO.TextWriter,Microsoft.Toolkit.Uwp.UI.Lottie.LottieData.Tools.Stats)")]
-[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.NamingRules", "SA1312:Variable names should begin with lower-case letter", Justification = "Rule is incorrect on anonymous tuple types.", Scope = "member", Target = "~M:Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen.TreeReducer.RemoveEmptyContainers(Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.ObjectGraph{Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen.TreeReducer.Node})")]
-[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.NamingRules", "SA1312:Variable names should begin with lower-case letter", Justification = "Rule is incorrect on anonymous tuple types.", Scope = "member", Target = "~M:Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen.TreeReducer.CoalesceContainerShapes(Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.ObjectGraph{Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen.TreeReducer.Node})")]
-[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.NamingRules", "SA1312:Variable names should begin with lower-case letter", Justification = "Rule is incorrect on anonymous tuple types.", Scope = "member", Target = "~M:Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen.TreeReducer.CoalesceContainerVisuals(Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.ObjectGraph{Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen.TreeReducer.Node})")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.NamingRules", "SA1312:Variable names should begin with lower-case letter", Justification = "Rule is incorrect on anonymous tuple types.", Scope = "member", Target = "~M:Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.TreeReducer.RemoveEmptyContainers(Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.ObjectGraph{Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.TreeReducer.Node})")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.NamingRules", "SA1312:Variable names should begin with lower-case letter", Justification = "Rule is incorrect on anonymous tuple types.", Scope = "member", Target = "~M:Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.TreeReducer.CoalesceContainerShapes(Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.ObjectGraph{Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.TreeReducer.Node})")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.NamingRules", "SA1312:Variable names should begin with lower-case letter", Justification = "Rule is incorrect on anonymous tuple types.", Scope = "member", Target = "~M:Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.TreeReducer.CoalesceContainerVisuals(Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.ObjectGraph{Microsoft.Toolkit.Uwp.UI.Lottie.UIData.Tools.TreeReducer.Node})")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Fix is too complex for now", Scope = "type", Target = "~T:Microsoft.Toolkit.Uwp.UI.Lottie.LottieData.Serialization.CheckedJsonObject")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Fix is too complex for now", Scope = "type", Target = "~T:Microsoft.Toolkit.Uwp.UI.Lottie.LottieData.Serialization.CheckedJsonArray")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "Fix is too complex for now", Scope = "type", Target = "~T:Microsoft.Toolkit.Uwp.UI.Lottie.LottieData.Serialization.JObjectExtensions")]
