+++ source/UIData/UIData.projitems
@@ -6,24 +6,12 @@
     <SharedGUID>74601E6C-2DFE-4842-B170-047941ABFF2C</SharedGUID>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\AlphanumericStringComparer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\CanonicalizedNode.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\Canonicalizer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\CodeBuilder.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\CppStringifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\CSharpInstantiatorGenerator.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\CxInstantiatorGenerator.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\IAnimatedVisualInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\IAnimatedVisualSourceInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\InstantiatorGeneratorBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\LoadedImageSurfaceInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\NodeNamer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\Optimizer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\Stringifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CodeGen\TreeReducer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Serialization\CompositionObjectDgmlSerializer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tools\CanonicalizedNode.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tools\Canonicalizer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Tools\Graph.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Tools\ObjectGraph.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tools\Optimizer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Tools\Stats.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Tools\TreeReducer.cs" />
   </ItemGroup>
 </Project>
