+++ source/UIDataCodeGen/CodeGen/AlphanumericStringComparer.cs
