+++ source/UIDataCodeGen/CodeGen/CppStringifier.cs
