+++ source/UIDataCodeGen/CodeGen/CxInstantiatorGenerator.cs
@@ -15,7 +15,7 @@ using Mgce = Microsoft.Toolkit.Uwp.UI.Lottie.WinCompData.Mgce;
 
 namespace Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen
 {
-#if PUBLIC_UIData
+#if PUBLIC_UIDataCodeGen
     public
 #endif
     sealed class CxInstantiatorGenerator : InstantiatorGeneratorBase
