+++ source/UIDataCodeGen/CodeGen/IAnimatedVisualSourceInfo.cs
@@ -10,7 +10,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen
     /// <summary>
     /// Holds information about code being generated for an IAnimatedVisualSource.
     /// </summary>
-#if PUBLIC_UIData
+#if PUBLIC_UIDataCodeGen
     public
 #endif
     interface IAnimatedVisualSourceInfo
