+++ source/UIDataCodeGen/Serialization/CompositionObjectDgmlSerializer.cs
